Added authentication, handle cookies for the token improve github authentication

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from 'zod';
 import { prisma } from '../lib/prisma';
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+    app.get('/memories', async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -19,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid()
@@ -32,6 +38,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id
             }
         })
+
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
 
         return memory
     })
@@ -50,14 +60,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'c5b91765-dda8-4c1b-b5b9-77a59dd32b75'
+                userId: request.user.sub
             }
         })
         return memory
 
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
@@ -72,8 +82,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id
+            }
+        })
 
-        const memory = await prisma.memory.update({
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -87,12 +106,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id
+            }
+        })
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
